refactor(chat): message_bubble god file를 content-type별 위젯으로 분할
1669줄짜리 message_bubble.dart를 thin dispatcher(887줄)로 줄이고
content-type별 위젯을 widgets/message_bubble/ 폴더로 추출:
- auto_download_image_widget.dart (이미지 버블, 자동 다운로드)
- dismissible_image_viewer.dart (전체화면 이미지 뷰어)
- forward_room_picker_dialog.dart (전달 대상 방 선택)
- message_bubble_parts.dart (system/failed/forwarded/reply/video/file 위젯)
- message_file_format.dart (파일 아이콘/크기 포맷 헬퍼)

동작·렌더링 보존(behavior-preserving): AppLocalizations 조회와 위젯 트리를
그대로 유지. golden 테스트(message_bubble_list.png) regen 없이 통과.
