+++ lib/presentation/pages/chat/widgets/message_bubble.dart
@@ -13,20 +13,27 @@ import '../../../../core/utils/save_image_to_gallery.dart';
 import '../../../../l10n/app_localizations.dart';
 import '../../../../domain/entities/message.dart';
 import '../../../../domain/entities/link_preview.dart';
-import '../../../../domain/entities/chat_room.dart';
 import '../../../blocs/chat/chat_room_bloc.dart';
 import '../../../blocs/chat/chat_room_event.dart';
 import '../../../blocs/chat/chat_list_bloc.dart';
-import '../../../blocs/chat/chat_list_state.dart';
 import '../../../blocs/settings/chat_settings_cubit.dart';
 import '../../../widgets/link_preview_card.dart';
 import '../../../widgets/link_preview_loader.dart';
 import '../../../widgets/reaction_display.dart';
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
-import 'video_player_page.dart';
+import 'message_bubble/auto_download_image_widget.dart';
+import 'message_bubble/dismissible_image_viewer.dart';
+import 'message_bubble/forward_room_picker_dialog.dart';
+import 'message_bubble/message_bubble_parts.dart';
 
 /// A message bubble widget that displays different types of messages.
 /// Handles text, image, and file messages with appropriate styling.
+///
+/// This is a thin dispatcher: per-content-type rendering lives in the
+/// `message_bubble/` folder (text/image/video/file/reply/forwarded/system/
+/// failed-status widgets), while the long-press action sheets and edit/delete/
+/// forward dialogs — which depend on this bubble's [message]/[isMe] and dispatch
+/// [ChatRoomBloc] events — remain here.
 class MessageBubble extends StatelessWidget {
   final Message message;
   final bool isMe;
@@ -40,191 +47,6 @@ class MessageBubble extends StatelessWidget {
   /// Check if edit time (5 minutes) has expired
   bool get _isEditTimeExpired {
     return DateTime.now().difference(message.createdAt).inMinutes >= 5;
-  }
-
-  /// Builds a reply preview shown above the message content
-  Widget _buildReplyPreview(BuildContext context) {
-    if (message.replyToMessageId == null) return const SizedBox.shrink();
-
-    // Try the embedded replyToMessage first, then look up from BLoC state
-    final replyMsg = message.replyToMessage ??
-        context.read<ChatRoomBloc>().state.messages
-            .where((m) => m.id == message.replyToMessageId)
-            .firstOrNull;
-
-    final l10n = AppLocalizations.of(context)!;
-    final previewText = replyMsg?.isDeleted == true
-        ? l10n.chatDeletedMessage
-        : (replyMsg?.content ?? l10n.chatOriginalMessageNotFound);
-    final senderName = replyMsg?.senderNickname ?? l10n.chatUnknownSender;
-
-    return Container(
-      margin: const EdgeInsets.only(bottom: 4),
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
-      constraints: BoxConstraints(
-        maxWidth: MediaQuery.of(context).size.width * 0.55,
-      ),
-      decoration: BoxDecoration(
-        color: isMe
-            ? Colors.white.withValues(alpha: 0.15)
-            : AppColors.primary.withValues(alpha: 0.08),
-        borderRadius: BorderRadius.circular(8),
-        border: Border(
-          left: BorderSide(
-            color: isMe ? Colors.white.withValues(alpha: 0.5) : AppColors.primary,
-            width: 2,
-          ),
-        ),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(
-            senderName,
-            style: TextStyle(
-              fontSize: 11,
-              fontWeight: FontWeight.w600,
-              color: isMe ? Colors.white.withValues(alpha: 0.9) : AppColors.primary,
-            ),
-          ),
-          Text(
-            previewText,
-            style: TextStyle(
-              fontSize: 12,
-              color: isMe
-                  ? Colors.white.withValues(alpha: 0.7)
-                  : context.textSecondaryColor,
-            ),
-            maxLines: 1,
-            overflow: TextOverflow.ellipsis,
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// Builds a "forwarded" indicator shown above the message
-  Widget _buildForwardedIndicator(BuildContext context) {
-    if (message.forwardedFromMessageId == null) return const SizedBox.shrink();
-
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 4),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(
-            Icons.forward,
-            size: 12,
-            color: isMe ? Colors.white.withValues(alpha: 0.6) : context.textSecondaryColor,
-          ),
-          const SizedBox(width: 4),
-          Text(
-            AppLocalizations.of(context)!.chatForwarded,
-            style: TextStyle(
-              fontSize: 11,
-              fontStyle: FontStyle.italic,
-              color: isMe ? Colors.white.withValues(alpha: 0.6) : context.textSecondaryColor,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  /// Builds failed status widget with retry and delete buttons
-  Widget _buildFailedStatusWidget(BuildContext context) {
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        // 전송 실패 아이콘
-        Icon(
-          Icons.error_outline,
-          size: 14,
-          color: Colors.red[400],
-        ),
-        const SizedBox(width: 4),
-        // 재전송 버튼
-        GestureDetector(
-          onTap: () {
-            if (message.localId != null) {
-              context.read<ChatRoomBloc>().add(MessageRetryRequested(message.localId!));
-            }
-          },
-          child: Container(
-            padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-            decoration: BoxDecoration(
-              color: AppColors.primary.withValues(alpha: 0.1),
-              borderRadius: BorderRadius.circular(4),
-            ),
-            child: Text(
-              AppLocalizations.of(context)!.chatResend,
-              style: const TextStyle(
-                color: AppColors.primary,
-                fontSize: 11,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(width: 4),
-        // 삭제 버튼
-        GestureDetector(
-          onTap: () {
-            if (message.localId != null) {
-              context.read<ChatRoomBloc>().add(PendingMessageDeleteRequested(message.localId!));
-            }
-          },
-          child: Container(
-            padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 2),
-            decoration: BoxDecoration(
-              color: Colors.red.withValues(alpha: 0.1),
-              borderRadius: BorderRadius.circular(4),
-            ),
-            child: Text(
-              AppLocalizations.of(context)!.commonDelete,
-              style: TextStyle(
-                color: Colors.red[400],
-                fontSize: 11,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  /// Returns appropriate icon for file type
-  IconData _getFileIcon(String? contentType) {
-    if (contentType == null) return Icons.insert_drive_file;
-    if (contentType.startsWith('image/')) return Icons.image;
-    if (contentType.startsWith('video/')) return Icons.videocam;
-    if (contentType.startsWith('audio/')) return Icons.audiotrack;
-    if (contentType.contains('pdf')) return Icons.picture_as_pdf;
-    if (contentType.contains('word') || contentType.contains('document')) {
-      return Icons.description;
-    }
-    if (contentType.contains('sheet') || contentType.contains('excel')) {
-      return Icons.table_chart;
-    }
-    if (contentType.contains('presentation') || contentType.contains('powerpoint')) {
-      return Icons.slideshow;
-    }
-    if (contentType.contains('zip') || contentType.contains('rar') || contentType.contains('tar')) {
-      return Icons.folder_zip;
-    }
-    return Icons.insert_drive_file;
-  }
-
-  /// Formats file size to human readable format
-  String _formatFileSize(int bytes) {
-    if (bytes < 1024) return '$bytes B';
-    if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
-    if (bytes < 1024 * 1024 * 1024) {
-      return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';
-    }
-    return '${(bytes / (1024 * 1024 * 1024)).toStringAsFixed(1)} GB';
   }
 
   /// Builds clickable text spans for URLs in message content
@@ -394,7 +216,7 @@ class MessageBubble extends StatelessWidget {
         opaque: false,
         barrierColor: Colors.black,
         pageBuilder: (context, animation, secondaryAnimation) {
-          return _DismissibleImageViewer(
+          return DismissibleImageViewer(
             imageUrl: imageUrl,
             heroTag: heroTag,
             onSaveToGallery: kIsWeb ? null : () => _saveImageToGallery(context, imageUrl),
@@ -509,7 +331,7 @@ class MessageBubble extends StatelessWidget {
       context: context,
       builder: (dialogContext) => BlocProvider.value(
         value: chatListBloc,
-        child: _ForwardRoomPickerDialog(
+        child: ForwardRoomPickerDialog(
           onRoomSelected: (roomId) {
             chatRoomBloc.add(MessageForwardRequested(
               messageId: message.id,
@@ -801,42 +623,15 @@ class MessageBubble extends StatelessWidget {
     );
   }
 
-  @override
-  Widget build(BuildContext context) {
-    // System messages are displayed centered
-    if (message.isSystemMessage) {
-      return Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-        child: Center(
-          child: Container(
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-            decoration: BoxDecoration(
-              color: context.dividerColor,
-              borderRadius: BorderRadius.circular(16),
-            ),
-            child: Text(
-              message.content,
-              style: TextStyle(
-                color: context.textSecondaryColor,
-                fontSize: 13,
-              ),
-              textAlign: TextAlign.center,
-            ),
-          ),
-        ),
-      );
-    }
-
-    // Time and status widget (카카오톡 스타일)
-    Widget timeWidget;
-
+  /// Builds the time + send-status widget shown beside the bubble.
+  Widget _buildTimeWidget(BuildContext context) {
     // 전송 실패 시: 재전송/삭제 버튼
     if (message.isFailed && isMe) {
-      timeWidget = _buildFailedStatusWidget(context);
+      return FailedStatusWidget(message: message);
     }
     // 전송 중: 로딩 인디케이터
-    else if (message.isPending && isMe) {
-      timeWidget = const SizedBox(
+    if (message.isPending && isMe) {
+      return const SizedBox(
         width: 12,
         height: 12,
         child: CircularProgressIndicator(
@@ -846,36 +641,34 @@ class MessageBubble extends StatelessWidget {
       );
     }
     // 전송 완료: 읽지 않음 개수 + 시간
-    else {
-      timeWidget = Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          if (message.unreadCount > 0 && isMe)
-            Padding(
-              padding: const EdgeInsets.only(right: 4),
-              child: Text(
-                '${message.unreadCount}',
-                style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                      color: AppColors.primary,
-                      fontWeight: FontWeight.w600,
-                      fontSize: 11,
-                    ),
-              ),
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        if (message.unreadCount > 0 && isMe)
+          Padding(
+            padding: const EdgeInsets.only(right: 4),
+            child: Text(
+              '${message.unreadCount}',
+              style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                    color: AppColors.primary,
+                    fontWeight: FontWeight.w600,
+                    fontSize: 11,
+                  ),
             ),
-          Text(
-            AppDateUtils.formatMessageTime(message.createdAt),
-            style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                  color: context.textSecondaryColor,
-                  fontSize: 11,
-                ),
           ),
-        ],
-      );
-    }
+        Text(
+          AppDateUtils.formatMessageTime(message.createdAt),
+          style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                color: context.textSecondaryColor,
+                fontSize: 11,
+              ),
+        ),
+      ],
+    );
+  }
 
-    // Message bubble content
-    Widget bubbleWidget;
-
+  /// Builds the content bubble for the message based on its type.
+  Widget _buildBubble(BuildContext context) {
     // Image message
     if (message.type == MessageType.image && message.fileUrl != null) {
       final imageUrl = message.fileUrl!;
@@ -892,7 +685,7 @@ class MessageBubble extends StatelessWidget {
       final chatSettings = context.read<ChatSettingsCubit>().state.settings;
       final autoDownload = chatSettings.autoDownloadImagesOnWifi; // Use wifi setting as the general toggle
 
-      bubbleWidget = _AutoDownloadImageWidget(
+      return AutoDownloadImageWidget(
         imageUrl: imageUrl,
         heroTag: heroTag,
         isMe: isMe,
@@ -902,252 +695,101 @@ class MessageBubble extends StatelessWidget {
       );
     }
     // Video message
-    else if (message.fileUrl != null && message.fileContentType?.startsWith('video/') == true) {
-      bubbleWidget = GestureDetector(
-        onTap: () {
-          Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (_) => VideoPlayerPage(
-                videoUrl: message.fileUrl!,
-                title: message.fileName,
-              ),
-            ),
-          );
-        },
-        child: ClipRRect(
-          borderRadius: BorderRadius.only(
-            topLeft: const Radius.circular(18),
-            topRight: const Radius.circular(18),
-            bottomLeft: Radius.circular(isMe ? 18 : 4),
-            bottomRight: Radius.circular(isMe ? 4 : 18),
-          ),
-          child: Container(
-            constraints: BoxConstraints(
-              maxWidth: MediaQuery.of(context).size.width * 0.65,
-            ),
-            width: 200,
-            height: 150,
-            color: Colors.black87,
-            child: Stack(
-              alignment: Alignment.center,
-              children: [
-                if (message.thumbnailUrl != null)
-                  Image.network(
-                    message.thumbnailUrl!,
-                    fit: BoxFit.cover,
-                    width: double.infinity,
-                    height: double.infinity,
-                    errorBuilder: (_, __, ___) => const SizedBox.shrink(),
-                  ),
-                Container(
-                  decoration: BoxDecoration(
-                    color: Colors.black.withValues(alpha: 0.3),
-                    shape: BoxShape.circle,
-                  ),
-                  padding: const EdgeInsets.all(12),
-                  child: const Icon(
-                    Icons.play_arrow,
-                    color: Colors.white,
-                    size: 32,
-                  ),
-                ),
-                // File name and size at the bottom
-                Positioned(
-                  bottom: 8,
-                  left: 8,
-                  right: 8,
-                  child: Row(
-                    children: [
-                      const Icon(Icons.videocam, color: Colors.white70, size: 14),
-                      const SizedBox(width: 4),
-                      Expanded(
-                        child: Text(
-                          message.fileName ?? AppLocalizations.of(context)!.chatVideo,
-                          style: const TextStyle(
-                            color: Colors.white70,
-                            fontSize: 11,
-                          ),
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      );
+    if (message.fileUrl != null && message.fileContentType?.startsWith('video/') == true) {
+      return VideoBubble(message: message, isMe: isMe);
     }
     // File message
-    else if (message.type == MessageType.file && message.fileUrl != null) {
-      bubbleWidget = GestureDetector(
+    if (message.type == MessageType.file && message.fileUrl != null) {
+      return FileBubble(
+        message: message,
+        isMe: isMe,
         onTap: () => _downloadFile(context, message.fileUrl!, message.fileName),
         onLongPress: () => _showFileOptions(context, message.fileUrl!, message.fileName),
-        child: Container(
-          constraints: BoxConstraints(
-            maxWidth: MediaQuery.of(context).size.width * 0.65,
-          ),
-          padding: const EdgeInsets.all(12),
-          decoration: BoxDecoration(
-            color: isMe ? context.myMessageBubbleColor : context.otherMessageBubbleColor,
-            borderRadius: BorderRadius.only(
-              topLeft: const Radius.circular(18),
-              topRight: const Radius.circular(18),
-              bottomLeft: Radius.circular(isMe ? 18 : 4),
-              bottomRight: Radius.circular(isMe ? 4 : 18),
-            ),
-            boxShadow: isMe
-                ? [
-                    BoxShadow(
-                      color: AppColors.primary.withValues(alpha: 0.15),
-                      blurRadius: 4,
-                      offset: const Offset(0, 2),
-                    ),
-                  ]
-                : [
-                    BoxShadow(
-                      color: Colors.black.withValues(alpha: 0.03),
-                      blurRadius: 2,
-                      offset: const Offset(0, 1),
-                    ),
-                  ],
-          ),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Container(
-                padding: const EdgeInsets.all(10),
-                decoration: BoxDecoration(
-                  color: isMe
-                      ? Colors.white.withValues(alpha: 0.2)
-                      : AppColors.primary.withValues(alpha: 0.1),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: Icon(
-                  _getFileIcon(message.fileContentType),
-                  color: isMe ? Colors.white : AppColors.primary,
-                  size: 24,
-                ),
-              ),
-              const SizedBox(width: 12),
-              Flexible(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      message.fileName ?? AppLocalizations.of(context)!.chatFileFallback,
-                      style: TextStyle(
-                        color: isMe ? Colors.white : context.textPrimaryColor,
-                        fontSize: 14,
-                        fontWeight: FontWeight.w500,
-                      ),
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    Row(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        if (message.fileSize != null)
-                          Text(
-                            _formatFileSize(message.fileSize!),
-                            style: TextStyle(
-                              color: isMe
-                                  ? Colors.white.withValues(alpha: 0.7)
-                                  : context.textSecondaryColor,
-                              fontSize: 12,
-                            ),
-                          ),
-                        const SizedBox(width: 8),
-                        Icon(
-                          Icons.download,
-                          size: 14,
-                          color: isMe
-                              ? Colors.white.withValues(alpha: 0.7)
-                              : context.textSecondaryColor,
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
       );
     }
     // Text message (default)
-    else {
-      final textColor = isMe ? Colors.white : context.textPrimaryColor;
+    final textColor = isMe ? Colors.white : context.textPrimaryColor;
 
-      // URL detection (show preview for first URL only)
-      final urlMatches = urlPattern.allMatches(message.displayContent);
-      final firstUrlRaw = urlMatches.isNotEmpty ? urlMatches.first.group(0) : null;
-      final firstUrl = firstUrlRaw != null ? normalizeUrl(firstUrlRaw) : null;
+    // URL detection (show preview for first URL only)
+    final urlMatches = urlPattern.allMatches(message.displayContent);
+    final firstUrlRaw = urlMatches.isNotEmpty ? urlMatches.first.group(0) : null;
+    final firstUrl = firstUrlRaw != null ? normalizeUrl(firstUrlRaw) : null;
 
-      bubbleWidget = Container(
-        constraints: BoxConstraints(
-          maxWidth: MediaQuery.of(context).size.width * 0.65,
+    return Container(
+      constraints: BoxConstraints(
+        maxWidth: MediaQuery.of(context).size.width * 0.65,
+      ),
+      padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 10),
+      decoration: BoxDecoration(
+        color: isMe ? context.myMessageBubbleColor : context.otherMessageBubbleColor,
+        borderRadius: BorderRadius.only(
+          topLeft: const Radius.circular(18),
+          topRight: const Radius.circular(18),
+          bottomLeft: Radius.circular(isMe ? 18 : 4),
+          bottomRight: Radius.circular(isMe ? 4 : 18),
         ),
-        padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 10),
-        decoration: BoxDecoration(
-          color: isMe ? context.myMessageBubbleColor : context.otherMessageBubbleColor,
-          borderRadius: BorderRadius.only(
-            topLeft: const Radius.circular(18),
-            topRight: const Radius.circular(18),
-            bottomLeft: Radius.circular(isMe ? 18 : 4),
-            bottomRight: Radius.circular(isMe ? 4 : 18),
-          ),
-          boxShadow: isMe
-              ? [
-                  BoxShadow(
-                    color: AppColors.primary.withValues(alpha: 0.15),
-                    blurRadius: 4,
-                    offset: const Offset(0, 2),
-                  ),
-                ]
-              : [
-                  BoxShadow(
-                    color: Colors.black.withValues(alpha: 0.03),
-                    blurRadius: 2,
-                    offset: const Offset(0, 1),
-                  ),
-                ],
-        ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            _buildForwardedIndicator(context),
-            _buildReplyPreview(context),
-            RichText(
-              text: TextSpan(
-                children: _buildTextSpans(context, message.displayContent, textColor),
-              ),
-            ),
-            // Link preview
-            if (message.hasLinkPreview)
-              LinkPreviewCard(
-                preview: LinkPreview(
-                  url: message.linkPreviewUrl!,
-                  title: message.linkPreviewTitle,
-                  description: message.linkPreviewDescription,
-                  imageUrl: message.linkPreviewImageUrl,
+        boxShadow: isMe
+            ? [
+                BoxShadow(
+                  color: AppColors.primary.withValues(alpha: 0.15),
+                  blurRadius: 4,
+                  offset: const Offset(0, 2),
                 ),
-                isMe: isMe,
-                maxWidth: MediaQuery.of(context).size.width * 0.6,
-              )
-            else if (firstUrl != null)
-              LinkPreviewLoader(
-                url: firstUrl,
-                isMe: isMe,
-                maxWidth: MediaQuery.of(context).size.width * 0.6,
+              ]
+            : [
+                BoxShadow(
+                  color: Colors.black.withValues(alpha: 0.03),
+                  blurRadius: 2,
+                  offset: const Offset(0, 1),
+                ),
+              ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          ForwardedIndicator(message: message, isMe: isMe),
+          ReplyPreview(message: message, isMe: isMe),
+          RichText(
+            text: TextSpan(
+              children: _buildTextSpans(context, message.displayContent, textColor),
+            ),
+          ),
+          // Link preview
+          if (message.hasLinkPreview)
+            LinkPreviewCard(
+              preview: LinkPreview(
+                url: message.linkPreviewUrl!,
+                title: message.linkPreviewTitle,
+                description: message.linkPreviewDescription,
+                imageUrl: message.linkPreviewImageUrl,
               ),
-          ],
-        ),
-      );
+              isMe: isMe,
+              maxWidth: MediaQuery.of(context).size.width * 0.6,
+            )
+          else if (firstUrl != null)
+            LinkPreviewLoader(
+              url: firstUrl,
+              isMe: isMe,
+              maxWidth: MediaQuery.of(context).size.width * 0.6,
+            ),
+        ],
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // System messages are displayed centered
+    if (message.isSystemMessage) {
+      return SystemMessageBubble(message: message);
     }
+
+    // Time and status widget (카카오톡 스타일)
+    final Widget timeWidget = _buildTimeWidget(context);
+
+    // Message bubble content
+    final Widget bubbleWidget = _buildBubble(context);
 
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
@@ -1240,430 +882,6 @@ class MessageBubble extends StatelessWidget {
           ),
         ],
       ),
-    );
-  }
-}
-
-/// 자동 다운로드 설정에 따라 이미지를 로드하는 위젯
-class _AutoDownloadImageWidget extends StatefulWidget {
-  final String imageUrl;
-  final String? heroTag;
-  final bool isMe;
-  final bool autoDownloadEnabled;
-  final VoidCallback onTapFullScreen;
-  final Function(BuildContext, String) onLongPress;
-
-  const _AutoDownloadImageWidget({
-    required this.imageUrl,
-    this.heroTag,
-    required this.isMe,
-    required this.autoDownloadEnabled,
-    required this.onTapFullScreen,
-    required this.onLongPress,
-  });
-
-  @override
-  State<_AutoDownloadImageWidget> createState() => _AutoDownloadImageWidgetState();
-}
-
-class _AutoDownloadImageWidgetState extends State<_AutoDownloadImageWidget> {
-  late bool _shouldLoad;
-
-  @override
-  void initState() {
-    super.initState();
-    _shouldLoad = widget.autoDownloadEnabled;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (!_shouldLoad) {
-      return GestureDetector(
-        onTap: () {
-          setState(() {
-            _shouldLoad = true;
-          });
-        },
-        child: ClipRRect(
-          borderRadius: BorderRadius.only(
-            topLeft: const Radius.circular(18),
-            topRight: const Radius.circular(18),
-            bottomLeft: Radius.circular(widget.isMe ? 18 : 4),
-            bottomRight: Radius.circular(widget.isMe ? 4 : 18),
-          ),
-          child: Container(
-            width: 200,
-            height: 150,
-            color: context.dividerColor,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Icon(
-                  Icons.image_outlined,
-                  color: context.textSecondaryColor,
-                  size: 40,
-                ),
-                const SizedBox(height: 8),
-                Text(
-                  AppLocalizations.of(context)!.chatTapToViewImage,
-                  style: TextStyle(
-                    color: context.textSecondaryColor,
-                    fontSize: 12,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      );
-    }
-
-    // 버블 썸네일은 화면상 maxWidth(= 화면폭 * 0.65) 로만 표시되므로,
-    // 디바이스 픽셀비를 곱한 적정 폭으로 memCacheWidth 를 지정해 서버 원본 해상도가
-    // 그대로 디코딩되어 메모리에 상주하는 것을 방지한다. (전체화면 뷰어는 풀해상도 유지)
-    final mq = MediaQuery.of(context);
-    final thumbCacheWidth =
-        (mq.size.width * 0.65 * mq.devicePixelRatio).round();
-
-    final cachedImage = CachedNetworkImage(
-      imageUrl: widget.imageUrl,
-      fit: BoxFit.cover,
-      memCacheWidth: thumbCacheWidth,
-      placeholder: (context, url) => Container(
-        width: 200,
-        height: 150,
-        color: context.dividerColor,
-        child: const Center(
-          child: CircularProgressIndicator(strokeWidth: 2),
-        ),
-      ),
-      errorWidget: (context, url, error) => Container(
-        width: 200,
-        height: 150,
-        color: context.dividerColor,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(Icons.broken_image, color: context.textSecondaryColor, size: 40),
-            const SizedBox(height: 8),
-            Text(
-              AppLocalizations.of(context)!.chatImageLoadFailed,
-              style: TextStyle(color: context.textSecondaryColor, fontSize: 12),
-            ),
-          ],
-        ),
-      ),
-    );
-
-    final borderRadius = BorderRadius.only(
-      topLeft: const Radius.circular(18),
-      topRight: const Radius.circular(18),
-      bottomLeft: Radius.circular(widget.isMe ? 18 : 4),
-      bottomRight: Radius.circular(widget.isMe ? 4 : 18),
-    );
-
-    // ClipRRect를 Hero 자식으로 넣어 Hero 비행 중에도 동일한 클리핑이 유지되도록 함.
-    // (ClipRRect를 Hero 바깥에 두면 비행 오버레이에서는 클리핑이 적용되지 않아 시각적 글리치 발생)
-    Widget imageChild = ClipRRect(
-      borderRadius: borderRadius,
-      child: Container(
-        constraints: BoxConstraints(
-          maxWidth: MediaQuery.of(context).size.width * 0.65,
-          maxHeight: 250,
-        ),
-        child: cachedImage,
-      ),
-    );
-
-    return GestureDetector(
-      onTap: widget.onTapFullScreen,
-      onLongPress: () => widget.onLongPress(context, widget.imageUrl),
-      child: widget.heroTag != null
-          ? Hero(
-              tag: widget.heroTag!,
-              // 비행 중에도 버블 측 둥근 모서리가 자연스럽게 유지됨
-              flightShuttleBuilder: (_, animation, direction, __, ___) {
-                return AnimatedBuilder(
-                  animation: animation,
-                  builder: (context, child) {
-                    // 출발(버블) → 도착(전체화면) 방향에 따라 radius를 보간
-                    final t = direction == HeroFlightDirection.push
-                        ? animation.value
-                        : 1 - animation.value;
-                    final radius = BorderRadius.lerp(
-                      borderRadius,
-                      BorderRadius.zero,
-                      t,
-                    )!;
-                    return ClipRRect(
-                      borderRadius: radius,
-                      child: Container(
-                        constraints: BoxConstraints(
-                          maxWidth: MediaQuery.of(context).size.width * 0.65,
-                          maxHeight: 250,
-                        ),
-                        child: cachedImage,
-                      ),
-                    );
-                  },
-                );
-              },
-              child: imageChild,
-            )
-          : imageChild,
-    );
-  }
-}
-
-/// 드래그로 닫을 수 있는 전체화면 이미지 뷰어 (카카오톡 스타일)
-class _DismissibleImageViewer extends StatefulWidget {
-  final String imageUrl;
-  final String? heroTag;
-  final VoidCallback? onSaveToGallery;
-
-  const _DismissibleImageViewer({
-    required this.imageUrl,
-    this.heroTag,
-    this.onSaveToGallery,
-  });
-
-  @override
-  State<_DismissibleImageViewer> createState() => _DismissibleImageViewerState();
-}
-
-class _DismissibleImageViewerState extends State<_DismissibleImageViewer>
-    with SingleTickerProviderStateMixin {
-  double _dragOffset = 0;
-  double _dragVelocity = 0;
-  late AnimationController _animationController;
-  late Animation<double> _animation;
-
-  // InteractiveViewer 줌 상태 추적용 컨트롤러.
-  // scale > 1.0 일 때만 panEnabled = true 로 설정해 드래그-투-디스미스와 충돌 방지.
-  final TransformationController _transformationController =
-      TransformationController();
-  bool _panEnabled = false;
-
-  static const double _dismissThreshold = 100;
-  static const double _velocityThreshold = 500;
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 200),
-    );
-    _animation = Tween<double>(begin: 0, end: 0).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.easeOut),
-    );
-    _animationController.addListener(() {
-      setState(() {
-        _dragOffset = _animation.value;
-      });
-    });
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    _transformationController.dispose();
-    super.dispose();
-  }
-
-  void _onInteractionUpdate(ScaleUpdateDetails details) {
-    // panEnabled 분기에 필요한 건 (scale > 1.0) boolean 전환 시점뿐이므로,
-    // scale 값 자체가 아니라 boolean 이 직전 값과 달라질 때만 setState 하여
-    // 줌 중 build/CachedNetworkImage 의 불필요한 리빌드를 줄인다.
-    final panEnabled =
-        _transformationController.value.getMaxScaleOnAxis() > 1.0;
-    if (panEnabled != _panEnabled) {
-      setState(() {
-        _panEnabled = panEnabled;
-      });
-    }
-  }
-
-  void _onVerticalDragUpdate(DragUpdateDetails details) {
-    setState(() {
-      _dragOffset += details.delta.dy;
-    });
-  }
-
-  void _onVerticalDragEnd(DragEndDetails details) {
-    _dragVelocity = details.velocity.pixelsPerSecond.dy;
-
-    final shouldDismiss = _dragOffset.abs() > _dismissThreshold ||
-        _dragVelocity.abs() > _velocityThreshold;
-
-    if (shouldDismiss) {
-      Navigator.of(context).pop();
-    } else {
-      // Snap back to center
-      _animation = Tween<double>(begin: _dragOffset, end: 0).animate(
-        CurvedAnimation(parent: _animationController, curve: Curves.easeOut),
-      );
-      _animationController.forward(from: 0);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final opacity = (1 - (_dragOffset.abs() / 300)).clamp(0.3, 1.0);
-    final scale = (1 - (_dragOffset.abs() / 1000)).clamp(0.8, 1.0);
-
-    return Scaffold(
-      backgroundColor: Colors.black.withValues(alpha: opacity),
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        iconTheme: IconThemeData(color: Colors.white.withValues(alpha: opacity)),
-        elevation: 0,
-        actions: [
-          if (widget.onSaveToGallery != null)
-            IconButton(
-              icon: Icon(Icons.download_rounded,
-                  color: Colors.white.withValues(alpha: opacity)),
-              tooltip: AppLocalizations.of(context)!.chatSaveToGallery,
-              onPressed: widget.onSaveToGallery,
-            ),
-        ],
-      ),
-      extendBodyBehindAppBar: true,
-      body: GestureDetector(
-        onVerticalDragUpdate: _onVerticalDragUpdate,
-        onVerticalDragEnd: _onVerticalDragEnd,
-        onTap: () => Navigator.of(context).pop(),
-        child: Container(
-          color: Colors.transparent,
-          child: Center(
-            child: Transform.translate(
-              offset: Offset(0, _dragOffset),
-              child: Transform.scale(
-                scale: scale,
-                child: InteractiveViewer(
-                  // scale == 1.0 일 때 pan 을 비활성화하여 드래그-투-디스미스 제스처가 동작하도록 함.
-                  // zoom 상태에서는 pan 활성화하여 정상 이동 가능.
-                  panEnabled: _panEnabled,
-                  transformationController: _transformationController,
-                  onInteractionUpdate: _onInteractionUpdate,
-                  minScale: 0.5,
-                  maxScale: 4,
-                  child: Builder(
-                    builder: (context) {
-                      final image = CachedNetworkImage(
-                        imageUrl: widget.imageUrl,
-                        fit: BoxFit.contain,
-                        placeholder: (context, url) => const Center(
-                          child: CircularProgressIndicator(
-                            color: Colors.white,
-                          ),
-                        ),
-                        errorWidget: (context, url, error) => const Center(
-                          child: Icon(Icons.broken_image, color: Colors.white54, size: 80),
-                        ),
-                      );
-                      if (widget.heroTag != null) {
-                        return Hero(tag: widget.heroTag!, child: image);
-                      }
-                      return image;
-                    },
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Dialog for selecting a chat room to forward a message to.
-class _ForwardRoomPickerDialog extends StatefulWidget {
-  final ValueChanged<int> onRoomSelected;
-
-  const _ForwardRoomPickerDialog({required this.onRoomSelected});
-
-  @override
-  State<_ForwardRoomPickerDialog> createState() => _ForwardRoomPickerDialogState();
-}
-
-class _ForwardRoomPickerDialogState extends State<_ForwardRoomPickerDialog> {
-  String _searchQuery = '';
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: Text(AppLocalizations.of(context)!.chatSelectRoom),
-      content: SizedBox(
-        width: double.maxFinite,
-        height: 400,
-        child: Column(
-          children: [
-            TextField(
-              decoration: InputDecoration(
-                hintText: AppLocalizations.of(context)!.chatSearchHint,
-                prefixIcon: const Icon(Icons.search),
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-                contentPadding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-              ),
-              onChanged: (value) => setState(() => _searchQuery = value.toLowerCase()),
-            ),
-            const SizedBox(height: 12),
-            Expanded(
-              child: BlocBuilder<ChatListBloc, ChatListState>(
-                builder: (context, state) {
-                  final rooms = state.chatRooms.where((room) {
-                    if (_searchQuery.isEmpty) return true;
-                    final name = room.displayName.toLowerCase();
-                    return name.contains(_searchQuery);
-                  }).toList();
-
-                  if (rooms.isEmpty) {
-                    return Center(child: Text(AppLocalizations.of(context)!.chatRoomListEmpty));
-                  }
-
-                  return ListView.builder(
-                    itemCount: rooms.length,
-                    itemBuilder: (context, index) {
-                      final room = rooms[index];
-                      return ListTile(
-                        leading: CircleAvatar(
-                          backgroundColor: AppColors.primaryLight,
-                          child: Icon(
-                            room.type == ChatRoomType.group
-                                ? Icons.group
-                                : Icons.person,
-                            color: Colors.white,
-                            size: 20,
-                          ),
-                        ),
-                        title: Text(
-                          room.displayName,
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                        ),
-                        onTap: () {
-                          Navigator.pop(context);
-                          widget.onRoomSelected(room.id);
-                        },
-                      );
-                    },
-                  );
-                },
-              ),
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: Text(AppLocalizations.of(context)!.commonCancel),
-        ),
-      ],
     );
   }
 }
